+++ src/tags/group-image.ts
@@ -9,7 +9,7 @@ const groupImage = (args: string[], content: string) => {
     let images: string[] = content.match(/<img[\s\S]*?>/g)!
     const rowCol = igRowCol(images.length)
     if (rowCol === 1) {
-        return imgElGlobalClass(images[0])
+        return imgElGlobalClass(images[0], ['img-swipe-single'])
     }
     if (imgsSameSize) {
         images = images.map((item) => {
@@ -23,7 +23,7 @@ const groupImage = (args: string[], content: string) => {
         })
     }
 
-    return `<div class="container group-image-container"><div class="row row-cols-${rowCol} gx-2 gy-2">${images.join(
+    return `<div class="container group-image-container img-swipe-group"><div class="row row-cols-${rowCol} gx-2 gy-2">${images.join(
         ''
     )}</div></div>`
 }