+++ src/utils/__test__/htmlEl.test.ts
@@ -6,4 +6,9 @@ describe('imgElGlobalClass', () => {
             '<img class="img-fluid object-fit-cover" alt="xxx" >'
         )
     })
+    test('classnames', () => {
+        expect(imgElGlobalClass('<img alt="xxx" >', ['aaa', 'bbb'])).toEqual(
+            '<img class="img-fluid object-fit-cover aaa bbb" alt="xxx" >'
+        )
+    })
 })