+++ src/utils/htmlEl.ts
@@ -1,3 +1,8 @@
-export function imgElGlobalClass(imgStr: string) {
-    return imgStr.replace('<img', '<img class="img-fluid object-fit-cover"')
+export function imgElGlobalClass(imgStr: string, classnames: string[] = []) {
+    return imgStr.replace(
+        '<img',
+        `<img class="${['img-fluid', 'object-fit-cover', ...classnames].join(
+            ' '
+        )}"`
+    )
 }